aplicado ajustes de seeder, e migration, criado recurso de login

+++ database/migrations/2022_10_26_210354_create_products_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->uuid('uuid')->unique();
             $table->string('name');
             $table->text('description');
-            $table->decimal('price', 8, 2);
+            $table->float('price', 8, 2);
             $table->unsignedInteger('stock');
             $table->timestamps();
         });
